adjust(naver-data): moved naverdata from services to NaverCard
