+++ src/components/Naver/index.ts
@@ -1,1 +1,2 @@
 export { default as NaverCard } from './NaverCard';
+export * from './NaverCard';
